Configure Schemes for Carthage
I made the ABControls scheme shared. This is necessary for Carthage to
be able to find and build the controls.

I duplicated the ABControls scheme and created the ABControls-Static
scheme. The static scheme will be used to generate the static framework
that can be linked directly into applications or other frameworks to
improve start times for applications.